@@ -18,12 +18,13 @@
     "url": "https://github.com/Scrounger/ioBroker.solarprognose.git"
   },
   "engines": {
-    "node": ">= 18"
+    "node": ">= 20"
   },
   "dependencies": {
-    "@iobroker/adapter-core": "^3.2.3",
-    "@types/node-schedule": "^2.1.7",
+    "@iobroker/adapter-core": "^3.3.2",
+    "lodash": "^4.17.21",
     "moment": "^2.30.1",
+    "undici": "^7.16.0",
     "node-schedule": "^2.1.1"
   },
   "devDependencies": {
@@ -31,30 +32,19 @@
     "@alcalzone/release-script-plugin-iobroker": "^3.7.2",
     "@alcalzone/release-script-plugin-license": "^3.7.0",
     "@alcalzone/release-script-plugin-manual-review": "^3.7.0",
-    "@iobroker/adapter-dev": "^1.3.0",
-    "@iobroker/testing": "^5.0.3",
-    "@tsconfig/node18": "^18.2.4",
-    "@types/chai": "^5.0.1",
-    "@types/chai-as-promised": "^8.0.1",
-    "@types/mocha": "^10.0.10",
-    "@types/node": "^22.13.1",
-    "@types/proxyquire": "^1.3.31",
-    "@types/sinon": "^17.0.3",
-    "@types/sinon-chai": "^4.0.0",
-    "@typescript-eslint/eslint-plugin": "^8.23.0",
-    "@typescript-eslint/parser": "^8.23.0",
-    "chai": "^5.1.2",
-    "chai-as-promised": "^8.0.1",
-    "eslint": "^9.20.0",
-    "mocha": "^11.1.0",
-    "proxyquire": "^2.1.3",
+    "@iobroker/adapter-dev": "^1.5.0",
+    "@iobroker/eslint-config": "^2.2.0",
+    "@iobroker/testing": "^5.1.1",
+    "@types/lodash": "^4.17.20",
+    "@types/node": "^24.5.2",
+    "@types/node-schedule": "^2.1.8",
     "rimraf": "^6.0.1",
-    "sinon": "^19.0.2",
-    "sinon-chai": "^4.0.0",
     "source-map-support": "^0.5.21",
     "ts-node": "^10.9.2",
-    "typescript": "~5.7.3"
+    "tsc-watch": "^7.1.1",
+    "typescript": "~5.9"
   },
+  "type": "module",
   "main": "build/main.js",
   "files": [
     "admin{,/!(src)/**}/!(tsconfig|tsconfig.*|.eslintrc).{json,json5}",
@@ -66,17 +56,18 @@
   ],
   "scripts": {
     "prebuild": "rimraf build",
-    "build": "build-adapter ts",
+    "build": "tsc",
+    "build-watch": "clear && tsc-watch --noClear --onSuccess 'iobroker restart solarprognose'",
     "watch": "build-adapter ts --watch",
     "prebuild:ts": "rimraf build",
-    "build:ts": "build-adapter ts",
-    "watch:ts": "build-adapter ts --watch",
+    "build:ts": "build-adapter ts --typescriptFormat esm",
+    "watch:ts": "build-adapter ts --watch --typescriptFormat esm",
     "test:ts": "mocha --config test/mocharc.custom.json src/**/*.test.ts",
     "test:package": "mocha test/package --exit",
     "test:integration": "mocha test/integration --exit",
     "test": "npm run test:ts && npm run test:package",
     "check": "tsc --noEmit",
-    "lint": "eslint --ext .ts src/",
+    "lint": "eslint -c eslint.config.mjs .",
     "translate": "translate-adapter",
     "release": "release-script"
   },
